pyproject.toml
[tool.poetry]
name = "sp-certification"
version = "0.1.0"
description = ""
authors = ["Valera K. <vk1337atliberty@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = ">=5.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
